+++ src/main/java/woojooin/planit/domain/goal/isa/dto/req/IsaAccountProductEditReq.java
@@ -1,4 +1,4 @@
-package woojooin.planit.domain.object.isa.dto.req;
+package woojooin.planit.domain.goal.isa.dto.req;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,7 +8,7 @@ import lombok.NonNull;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class IsaAccountProductRegisterReq {
+public class IsaAccountProductEditReq {
 
 	@NonNull
 	private Long goalId;
@@ -17,4 +17,7 @@ public class IsaAccountProductRegisterReq {
 	private Long memberProductId;
 
 	private String accountType = "ISA";
+
+	@NonNull
+	private boolean checked;
 }
